Added printing functionality to major classes.
Added print_state() function to Qubit() and Gate() classes. For Qubit() objects, the function outputs the state in Braket notation. For Gate() objects, the corresponding 2x2 matrix gets printed.

A bit more work has been done to the TensorProduct() class. The attribute self.state has been added to the TensorProduct(), which consists of a numpy array implementing the Kronecker product of all arguments. the corresponding print_state() function operates in one of two modes, which gets selected automatically. Braket-style printing is used for a vector (i.e. product of Qubit() objects), and the numpy array is printed to represent the matrix corresponding to a product of Gate() objects.